[fix] fix verbosity, null checking
- Respect the -v flag to get verbosity for tests
- Fix null checking so that it works for all datatypes --
previously we checked for type "S" but this does not exist.
Instead, use "O" for object and distinguish between string/bytes and
np.ndarray. Raise an error if the datatype is not as expected, to alert
that the test needs to be modified to handle a new datatype.
- Combine the test_non_null and test_all_beam_fields tests because they
are both loading all the data and are very slow, but basically test the
same things.